move sunrays component, cleanup, translations
