+++ src/screens/king-of-the-hill/KingOfTheHillExplainSheet.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useCallback, useMemo } from 'react';
+import React, { memo, useCallback } from 'react';
 import * as i18n from '@/languages';
 import { View, StyleSheet, Text as NativeText } from 'react-native';
 import { Panel, PANEL_WIDTH, TapToDismiss } from '@/components/SmoothPager/ListPanel';
@@ -8,24 +8,13 @@ import { SheetHandle } from '@/components/sheet';
 import { Box, Text, Separator, TextShadow, AnimatedText } from '@/design-system';
 import { foregroundColors, globalColors } from '@/design-system/color/palettes';
 import { THICK_BORDER_WIDTH } from '@/__swaps__/screens/Swap/constants';
-import { opacity } from '@/__swaps__/utils/swaps';
 import LinearGradient from 'react-native-linear-gradient';
 import { SmoothPager, usePagerNavigation } from '@/components/SmoothPager/SmoothPager';
 import { ButtonPressAnimation } from '@/components/animations';
 import { Extrapolation, interpolate, SharedValue, useDerivedValue } from 'react-native-reanimated';
 import { useNavigation } from '@/navigation';
 import chroma from 'chroma-js';
-import {
-  Canvas,
-  LinearGradient as SkiaLinearGradient,
-  RadialGradient,
-  Skia,
-  vec,
-  Circle,
-  Group,
-  Blur,
-  Path,
-} from '@shopify/react-native-skia';
+import { Canvas, LinearGradient as SkiaLinearGradient, RadialGradient, vec, Circle, Group, Blur } from '@shopify/react-native-skia';
 import { GradientText } from '@/components/text';
 import { AnimatedBlurView } from '@/components/AnimatedComponents/AnimatedBlurView';
 import { StepIndicators } from './components/StepInidicators';
@@ -33,6 +22,7 @@ import currentKingImage from '@/assets/kingOfTheHillExplainer/currentKing.png';
 import pointsMultiplierImage from '@/assets/kingOfTheHillExplainer/pointsMultiplier.png';
 import FastImage from 'react-native-fast-image';
 import { fonts } from '@/styles';
+import { Sunrays } from './components/Sunrays';
 
 const GRADIENT_COLORS = ['#8754C8', '#EE431D', '#FFF000', '#02ADDE'];
 const REVERSE_GRADIENT_COLORS = GRADIENT_COLORS.reverse();
@@ -42,40 +32,44 @@ const PANEL_HEADER_HEIGHT = 70;
 const PANEL_PADDING_HORIZONTAL = 14;
 const PANEL_INNER_WIDTH = PANEL_WIDTH - 2 * PANEL_PADDING_HORIZONTAL;
 
+const translations = i18n.l.king_of_hill.explain_sheet;
+const nextButtonLabel = i18n.t(translations.next);
+const gotItButtonLabel = i18n.t(translations.got_it);
+
 const STEPS = [
   {
     id: 'step-1',
-    title: 'The race is live',
+    title: i18n.t(translations.steps.step_1.title),
     graphicComponent: () => (
       <FastImage source={currentKingImage} style={{ width: PANEL_INNER_WIDTH, height: '100%' }} resizeMode={FastImage.resizeMode.contain} />
     ),
     subtitleComponent: () => (
       <Text align="center" size="17pt" weight="medium" color="labelTertiary" style={{ lineHeight: 22.95 }}>
-        {'The token with the most'}
+        {i18n.t(translations.steps.step_1.subtitle_parts[0])}
         <Text size="17pt" weight="bold" color="label">
-          {' buy volume '}
+          {i18n.t(translations.steps.step_1.subtitle_parts[1])}
         </Text>
-        {'right now is the current king.'}
+        {i18n.t(translations.steps.step_1.subtitle_parts[2])}
       </Text>
     ),
   },
   {
     id: 'step-2',
-    title: 'One crown a day',
+    title: i18n.t(translations.steps.step_2.title),
     graphicComponent: () => <NativeText style={{ fontSize: 90, fontFamily: fonts.family.SFProRounded, marginTop: -10 }}>{'👑'}</NativeText>,
     subtitleComponent: () => (
       <Text align="center" size="17pt" weight="medium" color="labelTertiary" style={{ lineHeight: 22.95 }}>
-        {'At'}
+        {i18n.t(translations.steps.step_2.subtitle_parts[0])}
         <Text size="17pt" weight="bold" color="label">
-          {' midnight (12:00 AM UTC), '}
+          {i18n.t(translations.steps.step_2.subtitle_parts[1])}
         </Text>
-        {'the leading token is crowned the daily winner.'}
+        {i18n.t(translations.steps.step_2.subtitle_parts[2])}
       </Text>
     ),
   },
   {
     id: 'step-3',
-    title: 'Win Rewards',
+    title: i18n.t(translations.steps.step_3.title),
     graphicComponent: () => (
       <FastImage
         source={pointsMultiplierImage}
@@ -86,119 +80,39 @@ const STEPS = [
     subtitleComponent: () => (
       <Box gap={16}>
         <Text align="center" size="17pt" weight="medium" color="labelTertiary">
-          {'⚡️ Buyers of the current King get'}
+          {`⚡️ ${i18n.t(translations.steps.step_3.subtitle_parts[0])}`}
           <Text size="17pt" weight="bold" color="label">
-            {' 2x points'}
+            {i18n.t(translations.steps.step_3.subtitle_parts[1])}
           </Text>
         </Text>
         <Text align="center" size="17pt" weight="medium" color="labelTertiary">
-          {'👑 Creators'}
+          {`👑 ${i18n.t(translations.steps.step_3.subtitle_two_parts[0])}`}
           <Text size="17pt" weight="bold" color="label">
-            {' earn points '}
+            {i18n.t(translations.steps.step_3.subtitle_two_parts[1])}
           </Text>
-          {'when crowned.'}
+          {i18n.t(translations.steps.step_3.subtitle_two_parts[2])}
         </Text>
       </Box>
     ),
   },
 ];
 
-function createConePath({ headWidth, baseWidth, height }: { headWidth: number; baseWidth: number; height: number }) {
-  const path = Skia.Path.Make();
-  path.moveTo((headWidth - baseWidth) / 2, height);
-  path.lineTo((headWidth + baseWidth) / 2, height);
-  path.lineTo(headWidth, 0);
-  path.lineTo(0, 0);
-  path.close();
+const BACKGROUND_CONFIG = {
+  glowCircleRadius: 100,
+  sunrayFocalSize: 85,
+  sunrayRayHeight: 136,
+  overflowBuffer: 100,
+};
 
-  return path;
-}
-
-function Sunrays({
-  rayCount,
-  rayFocalWidth,
-  rayHeadWidth,
-  rayHeight,
-  focalSize,
-  blur,
-}: {
-  rayCount: number;
-  rayFocalWidth: number;
-  rayHeadWidth: number;
-  rayHeight: number;
-  focalSize: number;
-  blur?: number;
-}) {
-  const focalRadius = focalSize / 2;
-  const size = 2 * (focalRadius + rayHeight);
-  const centerX = size / 2;
-  const centerY = size / 2;
-
-  const rayPath = useMemo(() => {
-    return createConePath({ headWidth: rayHeadWidth, baseWidth: rayFocalWidth, height: rayHeight });
-  }, [rayHeadWidth, rayFocalWidth, rayHeight]);
-
-  const rayTransforms = useMemo(() => {
-    // Create an array of x cones with equal angle increments
-    const rayAngles = Array.from({ length: rayCount }, (_, index) => ({
-      angle: (index * Math.PI * 2) / rayCount,
-    }));
-    return rayAngles.map(cone => {
-      // Calculate position on the circle
-      const x = centerX + focalRadius * Math.cos(cone.angle);
-      const y = centerY + focalRadius * Math.sin(cone.angle);
-
-      // Calculate the angle to point towards center (add 90° to align properly)
-      const pointToCenter = cone.angle + Math.PI / 2;
-
-      return [
-        { translateX: x },
-        { translateY: y },
-        // Rotate to point towards center
-        { rotate: pointToCenter },
-        // Move the cone so its base is at the circle point
-        { translateX: -rayHeadWidth / 2 },
-        { translateY: -rayHeight },
-      ];
-    });
-  }, [rayCount, centerX, focalRadius, centerY, rayHeadWidth, rayHeight]);
-
-  return (
-    <Group>
-      {rayTransforms.map((transform, index) => (
-        <Group key={index} transform={transform}>
-          <Path path={rayPath}></Path>
-          <SkiaLinearGradient
-            start={vec(rayHeadWidth / 2, 0)}
-            end={vec(rayHeadWidth / 2, rayHeight)}
-            colors={['rgba(255, 255, 255, 0)', 'rgba(255, 255, 255, 0.06)']}
-          />
-        </Group>
-      ))}
-      {blur && <Blur blur={blur} />}
-    </Group>
-  );
-}
-
-function SunraysBackground() {
-  const glowCircleRadius = 100;
-  const sunrayFocalSize = 85;
-  const sunrayRayHeight = 136;
+const SunraysBackground = memo(function SunraysBackground() {
+  const { glowCircleRadius, sunrayFocalSize, sunrayRayHeight, overflowBuffer } = BACKGROUND_CONFIG;
   const sunraysSize = sunrayRayHeight * 2 + sunrayFocalSize;
-  const overflowBuffer = 100;
 
   return (
     <Canvas style={[StyleSheet.absoluteFill, { marginHorizontal: -overflowBuffer }]}>
       <Group antiAlias dither transform={[{ translateX: overflowBuffer }]}>
         <Group transform={[{ translateX: (PANEL_WIDTH - sunraysSize) / 2 }, { translateY: PANEL_HEADER_HEIGHT / 2 }]}>
-          <Sunrays
-            rayCount={8}
-            rayFocalWidth={30}
-            rayHeadWidth={70}
-            rayHeight={sunrayRayHeight}
-            focalSize={sunrayFocalSize}
-            blur={6.88 / 2}
-          />
+          <Sunrays rayCount={8} rayFocalWidth={30} rayHeadWidth={70} rayHeight={sunrayRayHeight} focalSize={sunrayFocalSize} blur={4.44} />
         </Group>
         <Group transform={[{ translateX: PANEL_WIDTH / 2 - glowCircleRadius }, { translateY: PANEL_HEADER_HEIGHT + 36 }]}>
           <Circle blendMode={'plus'} cx={glowCircleRadius} cy={glowCircleRadius} r={glowCircleRadius}>
@@ -221,7 +135,7 @@ function SunraysBackground() {
       </Group>
     </Canvas>
   );
-}
+});
 
 const PanelBackground = memo(function PanelBackground() {
   return (
@@ -261,15 +175,15 @@ const PanelHeader = memo(function PanelHeader() {
           <Box style={StyleSheet.absoluteFill}>
             <GradientText colors={GRADIENT_COLORS} locations={[0, 0.5, 0.75, 1]} bleed={12}>
               <TextShadow shadowOpacity={1} blur={12}>
-                <Text size="20pt" weight="black" color="label" style={{ letterSpacing: 0.6 }}>
-                  {'KING OF THE HILL'}
+                <Text size="20pt" weight="black" color="label" uppercase style={{ letterSpacing: 0.6 }}>
+                  {i18n.t(i18n.l.king_of_hill.king_of_the_hill)}
                 </Text>
               </TextShadow>
             </GradientText>
           </Box>
           <GradientText colors={TEXT_GRADIENT_COLORS} locations={[0, 0.5, 0.75, 1]} bleed={12}>
-            <Text size="20pt" weight="black" color="label" style={{ letterSpacing: 0.6 }}>
-              {'KING OF THE HILL'}
+            <Text size="20pt" weight="black" color="label" uppercase style={{ letterSpacing: 0.6 }}>
+              {i18n.t(i18n.l.king_of_hill.king_of_the_hill)}
             </Text>
           </GradientText>
         </Box>
@@ -295,9 +209,9 @@ const Step = memo(function Step({
   });
 
   return (
-    <Box width={PANEL_WIDTH} justifyContent={'flex-end'} alignItems={'center'} style={{ flex: 1 }}>
+    <Box width={PANEL_WIDTH} justifyContent={'flex-end'} alignItems={'center'} style={styles.flex}>
       <SunraysBackground />
-      <Box alignItems={'center'} marginTop={{ custom: PANEL_HEADER_HEIGHT }} style={{ flex: 1 }}>
+      <Box alignItems={'center'} marginTop={{ custom: PANEL_HEADER_HEIGHT }} style={styles.flex}>
         <Box justifyContent={'center'} alignItems={'center'} width={PANEL_INNER_WIDTH} height={280}>
           {step.graphicComponent()}
         </Box>
@@ -332,7 +246,7 @@ const PanelContent = memo(function PanelContent() {
   });
 
   const buttonLabel = useDerivedValue(() => {
-    return (roundedCurrentPageIndex.value < STEPS.length - 1 ? 'Next' : 'Got it') as string;
+    return roundedCurrentPageIndex.value < STEPS.length - 1 ? nextButtonLabel : gotItButtonLabel;
   });
 
   const goToNextStepOrDismiss = useCallback(() => {
@@ -353,7 +267,7 @@ const PanelContent = memo(function PanelContent() {
       style={StyleSheet.absoluteFill}
       gap={32}
     >
-      <Box gap={20} style={{ flex: 1 }}>
+      <Box gap={20} style={styles.flex}>
         <SmoothPager enableSwipeToGoBack={true} enableSwipeToGoForward={'always'} initialPage={STEPS[0].id} ref={ref}>
           {STEPS.map((step, index) => (
             <SmoothPager.Page
@@ -398,5 +312,8 @@ const styles = StyleSheet.create({
     justifyContent: 'flex-end',
     paddingBottom: safeAreaInsetValues.bottom,
     pointerEvents: 'box-none',
+  },
+  flex: {
+    flex: 1,
   },
 });
